% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clocks.R
\name{cite_clock}
\alias{cite_clock}
\title{Cite a clock}
\usage{
cite_clock(clock_id, style = c("plain", "bibtex"))
}
\arguments{
\item{clock_id}{A clock identifier.}

\item{style}{\code{"plain"} or \code{"bibtex"}.}
}
\value{
A character string.
}
\description{
Cite a clock
}
\examples{
\dontrun{
cite_clock("horvath2013")
cat(cite_clock("dnamphenoage", "bibtex"))
}
}
