% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{coverage}
\alias{coverage}
\alias{coverage.falcon_result}
\title{Per-clock coverage and skip reasons}
\usage{
coverage(x)

\method{coverage}{falcon_result}(x)
}
\arguments{
\item{x}{A \code{falcon_result}.}
}
\value{
A data frame.
}
\description{
Per-clock coverage and skip reasons
}
\examples{
\dontrun{
coverage(res)
}
}
