% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{obs}
\alias{obs}
\alias{obs.falcon_data}
\alias{obs.falcon_result}
\title{Sample annotation}
\usage{
obs(x)

\method{obs}{falcon_data}(x)

\method{obs}{falcon_result}(x)
}
\arguments{
\item{x}{A \code{falcon_data} or \code{falcon_result}.}
}
\value{
A data frame, one row per sample.
}
\description{
Sample annotation
}
\examples{
\dontrun{
head(obs(d))
}
}
