% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_benchmark}
\alias{plot_benchmark}
\title{AA2 and AA1 counts per clock}
\usage{
plot_benchmark(bench)
}
\arguments{
\item{bench}{Output of \code{\link[=run_benchmark]{run_benchmark()}}.}
}
\value{
A ggplot.
}
\description{
AA2 and AA1 counts per clock
}
\examples{
\dontrun{
plot_benchmark(run_benchmark(res))
}
}
