% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_clock_chord}
\alias{plot_clock_chord}
\title{Circos chord diagram of CpG sharing between clocks}
\usage{
plot_clock_chord(x, min_shared = 5L, max_clocks = 24L)
}
\arguments{
\item{x}{A \code{falcon_result}.}

\item{min_shared}{Minimum shared CpGs for a chord to be drawn.}

\item{max_clocks}{Cap on ring size, for legibility.}
}
\value{
A ggplot.
}
\description{
The correlation heatmap says two clocks agree. This says how much of that
agreement is built in: chord width is the number of CpGs the two clocks have
literally in common. A pair with a thick chord and a high correlation has
told you much less than a pair with a high correlation and no chord.
}
\details{
Follows the circos convention of the single-cell aging literature: an outer
ring of colour-coded entities, chords in the interior weighted by the
strength of the relationship.
}
\examples{
\dontrun{
plot_clock_chord(res)
}
}
